Add Sudoku board component and related files
Implemented a new Sudoku board component with its HTML, SCSS, .ts and spec files. Created new interfaces and models to support the functionality of the board. Added routing for the board, along with a not-found component for handling 404s. Adjusted the indentation size in .editorconfig and updated app.component.html with a heading.